% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg_example.R
\name{ecg_example}
\alias{ecg_example}
\title{Get path to easyRecg example file}
\usage{
ecg_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If NULL, the example files will be listed.}
}
\description{
easyRecg comes with a few sample files in the inst/extdata directory.
This is a convenience function to access them
}
\examples{
ecg_example(path = NULL)
ecg_example(path = "muse/muse_ecg3.xml")
}
