% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse_strip_names.R
\name{muse_strip_names}
\alias{muse_strip_names}
\title{Remove patient names from a MUSE XML}
\usage{
muse_strip_names(file, output_file)
}
\arguments{
\item{file}{Input file name}

\item{output_file}{Output file name}
}
\value{
Silently returns \code{output_file}
}
\description{
Reads an XML file, removes patient first and last name, and writes out a new XML
}
