% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_muse_xml_directory.R
\name{read_muse_xml_directory}
\alias{read_muse_xml_directory}
\title{Read a directory of XML ECGs exported from MUSE. Returns an array of size (n_xmls, 5000, 12, 1).}
\usage{
read_muse_xml_directory(xml_directory)
}
\arguments{
\item{xml_directory}{Text string with the path to the xml files.}
}
\description{
Read a directory of XML ECGs exported from MUSE. Returns an array of size (n_xmls, 5000, 12, 1).
}
\examples{
\dontrun{
ecgs <- read_muse_xml_directory("data/muse_xml/")
}
}
