% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_muse_xml_ecg.R
\name{read_muse_xml_ecg}
\alias{read_muse_xml_ecg}
\title{Read a single XML representation of an ECG exported from MUSE. Returns a matrix with data for 12-leads}
\usage{
read_muse_xml_ecg(file, file_out = NA, numpyformat = TRUE)
}
\arguments{
\item{file}{Filename for the xml file that will be read}

\item{file_out}{Filename for exported data.  If NA, no file is written.  Only works if numpyformat is FALSE.}

\item{numpyformat}{Logical to determine if the array should be returned as a 5000x12 (FALSE) or a 1x5000 x 12 x 1 (TRUE, Default) format.  The latter is the standard input for AI-ECG algorithms}
}
\description{
Read a single XML representation of an ECG exported from MUSE. Returns a matrix with data for 12-leads
}
\examples{
\dontrun{
ecg_1 <- read_muse_xml_ecg("1_MUSSE_ecg.xml")
}
}
