% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_muse_xml_meta.R
\name{read_muse_xml_meta}
\alias{read_muse_xml_meta}
\title{Extract meta data from MUSE XML files}
\usage{
read_muse_xml_meta(file)
}
\arguments{
\item{file}{Either a path to a file or a collection of file paths}
}
\value{
A list of data frames (or tibbles)
}
\description{
Extract meta data from MUSE XML files
}
