% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_numpy.R
\name{read_numpy}
\alias{read_numpy}
\title{Read a numpy (python) data array into R}
\usage{
read_numpy(filename)
}
\arguments{
\item{filename}{Filename for the numpy (.npy) file}
}
\description{
Read a numpy (python) data array into R
}
\examples{
r_ecg <- read_numpy("digital_ecgs.npy")
}
