% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_numpy.R
\name{write_numpy}
\alias{write_numpy}
\title{Read a numpy (python) data array into R}
\usage{
write_numpy(rarrayname, filename)
}
\arguments{
\item{rarrayname}{The multidimensional array}

\item{filename}{Filename for the numpy (.npy) file that will be produced}
}
\description{
Read a numpy (python) data array into R
}
\examples{
\dontrun{
r_ecg <- read_numpy("digital_ecgs.npy")
}
}
